Store autocomplete data in ZSET only.
It might save space if we store only normalized strings in the
autocomplete ZSET. But it makes thing complicated: we have to lookup
an autocomplete entry in ZSET, then extract some key from it and then
again have to ask redis for the value of this extracted key (stored
separately).

We cannot rely on consistency (is therer such a key?), we have to know
how to extract a key from an ZSET entry, and everything is really
error-prone.

Instead we now store autocomplete entries and their "value" one after
another, separated by a separator string: mykey&&myvalue. I trust this
is more robust in the long run.