Special repo for redis-server install.
We need a more recent version of redis-server thatn is currently
provided by Ubuntu LTS. This can bite you as well in production: we
use redis-server versions >= 3.0 that support the ZRANGEBYLEX
command. This is needed at least for autocomplete functionality.

In production use it is sufficient to have a recent version of
redis-server running somewhere (older clients cope with newer servers,
apparently), so that `redis-cli` will connect to this server.