Add test helper to check enumerated choices.
In case of autocomplete (and other) sources we normally provide sources
with simple keys and complex values. For instance the keys are simple
numbers mapped to strings with umlauts, which denote the values.

The existing helper functions assume that we have keys similar to
values (which holds for simple external vocabs).

The additional "enum" test tries to reflect the real use cases better
than the existing ones.