Searching redis source now yields at most 10 items.
This way we can make sure, that the term searched is displayed in the
autocomplete widget selection, even if it is sorted wrong by the
widgets' javascript. With the former value of 100 returned terms it
could happen that the exact term was found but sorted down by
autocomplete widget so that the term did not appeared in first ten
items anymore and therefore was not pickable by the user at all.